chore(api): remove redundant generated test edits
Castiron-Internal-PR: https://github.com/openai/openai-python-internal/pull/51
Castiron-Source-SHA: 7a2d9c30234c7c63a3fcb266c173d426074d209f
Castiron-Public-Base-SHA: 5ac1e03e1ae8f08acf6213b3e44772734dbb8e59
