+++ .castiron.stats.yml
@@ -1,8 +1,8 @@
 schema_version: 1
-generation_id: 260c1f53-e64c-4d91-86a5-b85a85b57da0
+generation_id: d01088d9-a117-4981-9899-afab176a91ec
 openapi_spec_hash: a99ded1ea34cf528a9cd5f064167f26a
 openapi_transformed_spec_hash: e24c9d9339620c3cce8bbdd80e9ef8ed
 config_hash: 85382dd94c503b5d225adc7636a77c9f
-codegen_sha: 1a5da8d7cca1526c2c1fedc340b19b76dc234ab2
+codegen_sha: 160eee5e853d55c35205907a7b0bdd7151c74cec
 codegen_hash: 0e1cb892e3631438e55b55edd14899551f458be8d073e13d2c191730297562d6
-public_codegen_sha: 2bc1edac041ed7a11d9affc837ad05dd26eeb29d
+public_codegen_sha: 53459fc3cbb96f06dc69478d23b9703346a61c64
